fix: 변경된 멤버변수 반영 및 테스트파일명 변경
